Add more kong

+++ controller/bucket_delete.go
@@ -7,9 +7,12 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 )
 
-func (c *Controller) BucketDelete(bucket string) error {
+func (c *Controller) BucketDelete(bucket string, flagForce bool) error {
 	if bucket == "" {
 		return fmt.Errorf("no bucket name specified")
+	}
+	if !flagForce {
+		return fmt.Errorf("force flag required")
 	}
 	_, err := c.client.DeleteBucket(c.ctx, &s3.DeleteBucketInput{
 		Bucket: aws.String(bucket),
